Moved hooks to the hooks folder
